lss_cloudfunction/update_cf/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==1.9.0
google-cloud-pubsub==2.1.0